Successfully added metadata to mp3 and mp4.
There are still some quirks. For example the tag genre when working with
mp3 requires you to put the correct number corrispondence. I don't
really know why given that before this it was working.
Obv in m4a you need the genre name.

I'm however able to insert rDNS atoms that are recognized (with all the
other metadata) from AtomicParsley.