Node module available in the pkg/ directory.
It is possible to link it with `npm link` in the `pkg/` directory, and
`npm link lofty-rust` in a new npm project directory.

The file `pkg/package.json` is basically copied from Cargo.toml; the
file `pkg/index.js` is the entry point for the module, exporting the
function lofty() which accepts two parameters: an object containing the
metadata, and the file to work with; the file `pkg/lofty-wasm.wasm`
is the wasm-compiled rust source.

<rant>
I spent three hours trying to understand WHY the whole thing was not
working after creating a valid npm module.
Node was not spitting out any error, same thing for wasm. But If i tried
to compile main.rs and execute it manually, passing the correct
parameters, it was working just fine.

I spent three hours trying to debug that stupind error. Turns out I was
calling `JSON.stringify()` both in the lofty function in the npm module
and in the test project.

The rust JSON deserializer was accepting that thing without even
noticing. This might be a bug but I onestly don't care, it's past 1am.
</rant>