Add makefile script for packaging the module.
Created a makefile and a script to automatically create the package.json
file.

`cargo build` is actually called every time you call make, but that's a
bug for another day.

I could have used cargo-make but it isn't documented in the way i like
and it doesn't seem to be a way of specifying files as deps for a task.
I didn't try it tho.