Add the "proper" path of meta-writer.wasm
There isn't actually a proper way of finding the path of the file that
is currently being executed, just like __dirname does in that other type
of module. I hate node.

The path is now hardcoded. This will probably work flawlessly until the
end of time, but I hate it.

I also had to change the version to 0.1.2, because apparently deleting a
module in npm is bugged. I know it would have been like cheating, but I
wouldn't have needed it if 'npm link' worked.