Work around a segfault in wasi.start().
wasi.start() is going to throw a segfault when trying to access some
propertis, or when trying to return.

The segmentation fault will occour randomly, and I wasn't able to
reproduce it in a test (the issue occurred while executing feyr-js).

I'm now directly calling the exported _start() method. This means that
i'm not performing some checks and I'm not setting up (what I think is)
shared memory. But it should be safe for our use case.